created an object for array plot :)
